+++ main.cpp
@@ -1,25 +1,5 @@
-
-#pragma once
-
-
-class Noncopyable {
-protected:
-  // Object obj = other; 禁止
-  Noncopyable(const Noncopyable&) = delete;
-
-  // obj = other; 禁止
-  Noncopyable operator = (const Noncopyable&) = delete;
-
-  // コンストラクタ,デストラクタ default指定
-  Noncopyable() = default;
-  virtual ~Noncopyable() {};
-};
-
-
-// ここまでがNoncopyable
-//---------------------------------------------------------
-
 #include <iostream>
+#include "Noncopyable.h"
 
 class Object : public Noncopyable {
 private:
